Issue #17
Refactor the filename functions into a class with class methods.
