+++ landbosse/excelio/__init__.py
@@ -3,4 +3,4 @@ from .XlsxGenerator import XlsxGenerator
 from .XlsxManagerRunner import XlsxManagerRunner
 from .XlsxSerialManagerRunner import XlsxSerialManagerRunner
 from .XlsxParallelManagerRunner import XlsxParallelManagerRunner
-from .filename_functions import landbosse_input_dir
+from .XlsxFileOperations import XlsxFileOperations
